Update help command with new commands, fix circular dependency, fix more build issues
