+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "jest",
-    "build:watch": "npm run build -- --watch",
+    "build:watch": "npm run build:prod -- --watch",
     "build": "rimraf dist/ && tsc --project tsconfig.json",
     "build:prod": "rimraf dist/ && tsc --project tsconfig-build.json",
     "lint": "eslint '**/*.ts' --quiet --fix",
