+++ src/types/price-patterns.ts
@@ -1,0 +1,7 @@
+export enum PricePatterns {
+    fluctuating = 'fluctuating',
+    decreasing = 'decreasing',
+    largeSpike = 'large spike',
+    smallSpike = 'small spike',
+    unknown = 'unknown',
+}
